Update to MUSCLE 5.72

+++ unizone/src/muscle/util/Directory.cpp
@@ -147,9 +147,9 @@ status_t Directory :: SetDir(const char * dirPath)
    Reset();
    if (dirPath)
    {
-      int pathLen = strlen(dirPath);
+      int pathLen = (int) strlen(dirPath);
       const char * sep = GetFilePathSeparator();
-      int sepLen = strlen(sep);
+      int sepLen = (int) strlen(sep);
       int extraBytes = ((pathLen<sepLen)||(strcmp(dirPath+pathLen-sepLen, sep) != 0)) ? sepLen : 0;
       _path = newnothrow_array(char, pathLen+extraBytes+1);
       if (_path == NULL) {WARN_OUT_OF_MEMORY; return B_ERROR;}
@@ -177,7 +177,7 @@ status_t Directory :: MakeDirectory(const char * dirPath, bool forceCreateParent
       const char * lastSlash = strrchr(dirPath+((dirPath[0]==sep)?1:0), sep);
       if (lastSlash)
       {
-         uint32 subLen = lastSlash-dirPath;
+         uint32 subLen = (uint32)(lastSlash-dirPath);
          char * temp = newnothrow_array(char, subLen+1);
          if (temp == NULL) {WARN_OUT_OF_MEMORY; return B_ERROR;}
 
@@ -204,7 +204,7 @@ status_t Directory :: MakeDirectory(const char * dirPath, bool forceCreateParent
 
 status_t Directory :: MakeDirectoryForFile(const char * filePath)
 {
-   int pathLen = strlen(filePath);
+   int pathLen = (int) strlen(filePath);
    char * p = newnothrow_array(char,pathLen+1);
    if (p == NULL) {WARN_OUT_OF_MEMORY; return B_ERROR;}
 
@@ -224,8 +224,8 @@ status_t Directory :: DeleteDirectory(const char * dirPath, bool forceDeleteSubI
       if (d.SetDir(dirPath) != B_NO_ERROR) return B_ERROR;
 
       const char * sep = GetFilePathSeparator();
-      int sepLen       = strlen(sep);
-      int dirPathLen   = strlen(dirPath);
+      int sepLen       = (int) strlen(sep);
+      int dirPathLen   = (int) strlen(dirPath);
 
       // No point in including a separator if (dirPath) already ends in one
       if ((dirPathLen >= sepLen)&&(strcmp(&dirPath[dirPathLen-sepLen], sep) == 0)) {sep = ""; sepLen=0;}
@@ -234,7 +234,7 @@ status_t Directory :: DeleteDirectory(const char * dirPath, bool forceDeleteSubI
       {
          if ((strcmp(fn, ".") != 0)&&(strcmp(fn, "..") != 0))
          {
-            int fnLen = strlen(fn);
+            int fnLen = (int) strlen(fn);
             char * catStr = newnothrow_array(char, dirPathLen+sepLen+fnLen+1);
             if (catStr == NULL) {WARN_OUT_OF_MEMORY; return B_ERROR;}
